Move commands/util.py to util/
