+++ mautrix_telegram/util/__init__.py
@@ -1,1 +1,2 @@
 from .file_transfer import transfer_file_to_matrix
+from .format_duration import format_duration
