+++ mautrix_telegram/util/format_duration.py
